LayoutTests/imported/w3c/web-platform-tests/svg/struct/scripted/svg-checkIntersection-002-expected.txt


FAIL checkIntersection() assert_equals: checkIntersection(), element g_inner_top rect 10,10,80,10 expected true but got false
